Rename Win32Device sources to snake_case
