+++ include/game_engine_device/miles_audio_device/miles_audio_manager.h
@@ -16,7 +16,7 @@
 **	along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-// FILE: MilesAudioManager.h //////////////////////////////////////////////////////////////////////////
+// FILE: miles_audio_manager.h //////////////////////////////////////////////////////////////////////////
 // MilesAudioManager implementation
 // Author: John K. McDonald, July 2002
