+++ include/game_engine_device/w3d_device/game_client/w3dmouse.h
@@ -34,7 +34,7 @@
 //
 // Project:    RTS3
 //
-// File name:  Win32Mouse.h
+// File name:  win32_mouse.h
 //
 // Created:    Mark Wilczynski, Jan 2002
 //
@@ -51,7 +51,7 @@
 // SYSTEM INCLUDES ////////////////////////////////////////////////////////////
 
 // USER INCLUDES //////////////////////////////////////////////////////////////
-#include "Win32Device/game_client/Win32Mouse.h"
+#include "win32_device/game_client/win32_mouse.h"
 
 // FORWARD REFERENCES /////////////////////////////////////////////////////////
 class CameraClass;
