+++ include/game_engine_device/win32_device/common/win32_game_engine.h
@@ -38,9 +38,9 @@
 #include "common/GameEngine.h"
 #include "game_logic/game_logic.h"
 #include "GameNetwork/NetworkInterface.h"
-#include "MilesAudioDevice/MilesAudioManager.h"
-#include "Win32Device/common/Win32BIGFileSystem.h"
-#include "Win32Device/common/Win32LocalFileSystem.h"
+#include "miles_audio_device/miles_audio_manager.h"
+#include "win32_device/common/Win32BIGFileSystem.h"
+#include "win32_device/common/Win32LocalFileSystem.h"
 #include "w3d_device/common/W3DModuleFactory.h"
 #include "w3d_device/game_logic/W3Dgame_logic.h"
 #include "w3d_device/game_client/W3Dgame_client.h"
