+++ include/game_engine_device/win32_device/game_client/win32_di_keyboard.h
@@ -16,7 +16,13 @@
 **	along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-// FILE: Win32DIMouse.h ///////////////////////////////////////////////////////
+////////////////////////////////////////////////////////////////////////////////
+//																																						//
+//  (c) 2001-2003 Electronic Arts Inc.																				//
+//																																						//
+////////////////////////////////////////////////////////////////////////////////
+
+// FILE: win32_di_keyboard.h ////////////////////////////////////////////////////
 //-----------------------------------------------------------------------------
 //
 //                       Westwood Studios Pacific.
@@ -28,19 +34,20 @@
 //
 // Project:    RTS3
 //
-// File name:  Win32DIMouse.h
+// File name:  win32_di_keyboard.h
 //
 // Created:    Colin Day, June 2001
 //
-// Desc:       Win32 direct input implementation for the mouse
+// Desc:       Device implementation of the keyboard interface on Win32
+//						 using Microsoft Direct Input
 //
 //-----------------------------------------------------------------------------
 ///////////////////////////////////////////////////////////////////////////////
 
 #pragma once
 
-#ifndef __WIN32DIMOUSE_H_
-#define __WIN32DIMOUSE_H_
+#ifndef __WIN32DIKEYBOARD_H_
+#define __WIN32DIKEYBOARD_H_
 
 // SYSTEM INCLUDES ////////////////////////////////////////////////////////////
 #ifndef DIRECTINPUT_VERSION
@@ -50,53 +57,47 @@
 #include <dinput.h>
 
 // USER INCLUDES //////////////////////////////////////////////////////////////
-#include "game_client/Mouse.h"
+#include "game_client/Keyboard.h"
 
 // FORWARD REFERENCES /////////////////////////////////////////////////////////
 
 // TYPE DEFINES ///////////////////////////////////////////////////////////////
 
-// class DirectInputMouse -----------------------------------------------------
-/** Direct input implementation for the mouse device */
+// class DirectInputKeyboard --------------------------------------------------
+/** Class for interfacing with the keyboard using direct input as the
+ * implementation */
 //-----------------------------------------------------------------------------
-class DirectInputMouse : public Mouse
+class DirectInputKeyboard : public Keyboard
 {
 
 public:
-	DirectInputMouse(void);
-	virtual ~DirectInputMouse(void);
+	DirectInputKeyboard(void);
+	virtual ~DirectInputKeyboard(void);
 
-	// extended methods from base class
-	virtual void init(void);				///< initialize the direct input mouse, extending functionality
-	virtual void reset(void);				///< reset system
-	virtual void update(void);				///< update the mouse data, extending functionality
-	virtual void setPosition(Int x, Int y); ///< set position for mouse
-
-	virtual void setMouseLimits(void); ///< update the limit extents the mouse can move in
-
-	virtual void setCursor(MouseCursor cursor); ///< set mouse cursor
-
-	virtual void capture(void);		   ///< capture the mouse
-	virtual void releaseCapture(void); ///< release mouse capture
+	// extend methods from the base class
+	virtual void init(void);		 ///< initialize the keyboard, extending init functionality
+	virtual void reset(void);		 ///< Reset the keybaord system
+	virtual void update(void);		 ///< update call, extending update functionality
+	virtual Bool getCapsState(void); ///< get state of caps lock key, return TRUE if down
 
 protected:
-	/// device implementation to get mouse event
-	virtual UnsignedByte getMouseEvent(MouseIO *result, Bool flush);
+	// extended methods from the base class
+	virtual void getKey(KeyboardIO *key); ///< get a single key event
 
-	// new internal methods for our direct input implemetation
-	void openMouse(void);  ///< create the direct input mouse
-	void closeMouse(void); ///< close and release mouse resources
-	/// map direct input mouse data to our own format
-	void mapDirectInputMouse(MouseIO *mouse, DIDEVICEOBJECTDATA *mdat);
+	//-----------------------------------------------------------------------------------------------
 
-	// internal data members for our direct input mouse
-	LPDIRECTINPUT8 m_pDirectInput;		 ///< pointer to direct input interface
-	LPDIRECTINPUTDEVICE8 m_pMouseDevice; ///< pointer to mouse device
+	// new methods to this derived class
+	void openKeyboard(void);  ///< create direct input keyboard
+	void closeKeyboard(void); ///< release direct input keyboard
 
-}; // end class DirectInputMouse
+	// direct input data members
+	LPDIRECTINPUT8 m_pDirectInput;			///< pointer to direct input interface
+	LPDIRECTINPUTDEVICE8 m_pKeyboardDevice; ///< pointer to keyboard device
+
+}; // end class DirectInputKeyboard
 
 // INLINING ///////////////////////////////////////////////////////////////////
 
 // EXTERNALS //////////////////////////////////////////////////////////////////
 
-#endif // __WIN32DIMOUSE_H_
+#endif // __WIN32DIKEYBOARD_H_
