+++ include/game_engine_device/win32_device/game_client/win32_mouse.h
@@ -22,7 +22,7 @@
 //																																						//
 ////////////////////////////////////////////////////////////////////////////////
 
-// FILE: Win32Mouse.h /////////////////////////////////////////////////////////
+// FILE: win32_mouse.h /////////////////////////////////////////////////////////
 //-----------------------------------------------------------------------------
 //
 //                       Westwood Studios Pacific.
@@ -34,7 +34,7 @@
 //
 // Project:    RTS3
 //
-// File name:  Win32Mouse.h
+// File name:  win32_mouse.h
 //
 // Created:    Colin Day, July 2001
 //
