+++ src/game_engine_device/CMakeLists.txt
@@ -24,35 +24,35 @@ file(GLOB W3DDEVICE_GAMELOGIC_SRCS
 )
 target_sources(gameenginedevice PRIVATE ${W3DDEVICE_GAMELOGIC_SRCS})
 
-# Win32Device sources migrated from Generals/Code
+# win32_device sources migrated from Generals/Code
 file(GLOB_RECURSE WIN32DEVICE_COMMON_SRCS CONFIGURE_DEPENDS
-    Win32Device/common/*.cpp
+    win32_device/common/*.cpp
 )
 file(GLOB_RECURSE WIN32DEVICE_GAMECLIENT_SRCS CONFIGURE_DEPENDS
-    Win32Device/GameClient/*.cpp
+    win32_device/game_client/*.cpp
 )
 target_sources(gameenginedevice PRIVATE
     ${WIN32DEVICE_COMMON_SRCS}
     ${WIN32DEVICE_GAMECLIENT_SRCS}
 )
 
-# MilesAudioDevice sources
+# miles_audio_device sources
 file(GLOB MILESAUDIODEVICE_SRCS
-    MilesAudioDevice/*.cpp
+    miles_audio_device/*.cpp
 )
 target_sources(gameenginedevice PRIVATE ${MILESAUDIODEVICE_SRCS})
 
 # VideoDevice sources
 file(GLOB_RECURSE VIDEODEVICE_SRCS CONFIGURE_DEPENDS
-    VideoDevice/*.cpp
+    video_device/*.cpp
 )
 target_sources(gameenginedevice PRIVATE ${VIDEODEVICE_SRCS})
 
 target_include_directories(gameenginedevice PUBLIC
     ${PROJECT_SOURCE_DIR}/include
-    ${PROJECT_SOURCE_DIR}/include/GameEngine
-    ${PROJECT_SOURCE_DIR}/include/GameEngineDevice
-    ${PROJECT_SOURCE_DIR}/include/Precompiled
+    ${PROJECT_SOURCE_DIR}/include/game_engine
+    ${PROJECT_SOURCE_DIR}/include/game_engine_device
+    ${PROJECT_SOURCE_DIR}/include/pre_compiled
 )
 
 # Link against the Miles Sound System stub which in turn
