+++ src/game_engine_device/lvgl_device/common/lvgl_big_file.cpp
@@ -4,7 +4,7 @@
 #include "common/StreamingArchiveFile.h"
 #include "common/GameMemory.h"
 #include "common/PerfTimer.h"
-#include "game_engine_device/lvgl_device/common/win32bigfile.h"
+#include "game_engine_device/lvgl_device/common/win32_big_file.h"
 
 Win32BIGFile::Win32BIGFile() {
 }
