+++ src/game_engine_device/win32_device/common/win32_cd_manager.cpp
