+++ src/game_engine_device/win32_device/common/win32_local_file.cpp
@@ -26,7 +26,7 @@
 // Bryan Cleveland, August 2002
 ///////////////////////////////////////////////////////
 
-#include "Win32Device/common/Win32LocalFile.h"
+#include "win32_device/common/Win32LocalFile.h"
 
 Win32LocalFile::Win32LocalFile() : LocalFile() {
 }
