+++ src/tools/GUIEdit/Source/GUIEdit.cpp
@@ -82,10 +82,10 @@
 #include "W3DDevice/GameClient/W3DGameFont.h"
 #include "W3DDevice/GameClient/W3DDisplayStringManager.h"
 #include "GameClient/Keyboard.h"
-#include "Win32Device/GameClient/Win32Mouse.h"
-#include "Win32Device/GameClient/Win32DIKeyboard.h"
-#include "Win32Device/common/win32localfilesystem.h"
-#include "Win32Device/common/win32bigfilesystem.h"
+#include "win32_device/game_client/win32_mouse.h"
+#include "win32_device/game_client/win32_di_keyboard.h"
+#include "win32_device/common/win32_local_file_system.h"
+#include "win32_device/common/win32_big_file_system.h"
 
 #include "Resource.h"
 #include "WinMain.h"
